requirements.txt
flask
mysql-connector-python
py-bcrypt